🧪 Add comprehensive tests for PlaybackQueueHelpers
